requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
discord.py==1.7.1
idna==2.10
multidict==5.1.0
praw==7.2.0
prawcore==2.0.0
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.4
uwuify==1.0.1
websocket-client==0.58.0
yarl==1.6.3
